MscmpSystInstance.Application behaviour (mscmp_syst_instance v0.1.0)

Defines a behaviour and provides support functions for MscmpSystInstance compatible Application modules.

Callbacks

child_spec(opts)

@callback child_spec(opts :: Keyword.t()) :: Supervisor.child_spec()

Returns the child specification of the Application.

Parameters

opts - Miscellaneous options which the specific Application implementation of this Behaviour may define for it's own purposes.

create_instance(instance_name, startup_options, opts)

@callback create_instance(
  instance_name :: MscmpSystInstance.Types.instance_name(),
  startup_options :: MscmpSystOptions.Types.options(),
  opts :: Keyword.t()
) :: :ok | {:error, Exception.t()}

Creates a new Application Instance.

Parameters

instance_name - The name of the Instance to be created. This name must be globally unique amongst Instance Names, including Instances of other Applications. This value is required.

startup_options - The startup options which provide database connection information for where the Instance and its contexts are to be created. See MscmpSystOptions for more. This value is required.

opts - Miscellaneous options which the specific Application implementation of this Behaviour may define for it's own purposes.

delete_instance(instance_name, startup_options, opts)

@callback delete_instance(
  instance_name :: MscmpSystInstance.Types.instance_name(),
  startup_options :: MscmpSystOptions.Types.options(),
  opts :: Keyword.t()
) :: :ok | {:error, Exception.t()}

Deletes an existing Application Instance.

Note that specific Application implementation may impose requirements such as Instances to be deleted must be stopped prior to deletion or that they might have attributes such as "purge elegible" flags which must be set prior to deletion. As a behaviour, this callback does not hold any opinions on such matters.

Parameters

instance_name - The name of the Instance to be created. This name must be globally unique amongst Instance Names, including Instances of other Applications. This value is required.

startup_options - The startup options which provide database connection information for the Instance and its contexts which are to be deleted. See MscmpSystOptions for more. This value is required.

opts - Miscellaneous options which the specific Application implementation of this Behaviour may define for it's own purposes.

get_instance_services(instance_name)

@callback get_instance_services(instance_name :: MscmpSystInstance.Types.instance_name()) ::
  MscmpSystInstance.Types.services()

Retrieves a mapping of the known Instance services mapped to a specific Instance's runtime identifiers for those services.

Parameters

instance_name - the Instance for which to return service mappings.

get_instance_states(instances)

@callback get_instance_states(instances :: MscmpSystInstance.Types.startable_instances()) ::
  [
    MscmpSystInstance.Types.instance_state()
  ]

Retrieves the lifecycle and runtime states of the requested Application Instances.

Parameters

instances - the desired Instances for which to return the runtime state.

get_services()

@callback get_services() :: MscmpSystInstance.Types.services()

Returns the map of Application services to their runtime names or PIDs.

An Application may optionally support one or more runtime services, typically for the benefit of its own Instances; these might include an Application level Registry or PubSub services.

init(instances, startup_options, opts)

@callback init(
  instances :: MscmpSystInstance.Types.instances(),
  startup_options :: MscmpSystOptions.Types.options(),
  opts :: Keyword.t()
) ::
  {:ok, pid()}
  | {:error, {:already_started, pid()} | {:shutdown, term()} | term()}

Initializes the Application, its services, and optionally starts its Instances.

Parameters

instances - Identifies the Application Instances to start, if any.

startup_options - The startup options to use when starting the application. Typically these options will include externally configured runtime options such as database connection information. See MscmpSystOptions for more.

opts - Additional options to pass to the Application start process. The specific options available will vary by Application implementation.

put_instance(instance_name)

@callback put_instance(instance_name :: MscmpSystInstance.Types.instance_name()) ::
  MscmpSystInstance.Types.instance_name() | nil

Establishes the current Instance services which should be used for the current process.

As Instance may start a number of services, including Datastore Contexts (database connections), PubSub services, and so on. This function will set the current the map of services to use for the process in the Process Dictionary so that any calls to those services will be in the context of the "put" Instance's services.

Parameters

instance_name - the Instance which will service calls for any business logic invocation of Instance services.

Returns

The return value is that of the previous Instance set or nil if no previous Instance was set.

start_instances(instances, startup_options, opts)

@callback start_instances(
  instances :: MscmpSystInstance.Types.startable_instances(),
  startup_options :: MscmpSystOptions.Types.options(),
  opts :: Keyword.t()
) ::
  {:ok, [MscmpSystInstance.Types.instance_action_result()]}
  | {:error, Exception.t()}

Starts unstarted, but startable, Application Instances.

Parameters

instances - Identifies the Application Instances to start.

startup_options - The startup options to use when starting the application. Typically these options will include externally configured runtime options such as database connection information. See MscmpSystOptions for more.

stop()

@callback stop() :: :ok

Stops the Application services and any of its running Instances.

stop_instances(instances)

@callback stop_instances(instances :: MscmpSystInstance.Types.startable_instances()) ::
  {:ok, [MscmpSystInstance.Types.instance_action_result()]}
  | {:error, Exception.t()}

Stops the identified running Application Instances.

Parameters

instances - Identifies the Application Instances to stop.

upgrade_instances(instances, startup_options, opts)

@callback upgrade_instances(
  instances :: MscmpSystInstance.Types.startable_instances(),
  startup_options :: MscmpSystOptions.Types.options(),
  opts :: Keyword.t()
) ::
  {:ok, [MscmpSystInstance.Types.instance_action_result()]}
  | {:error, Exception.t()}

Upgrades the identified Application Instances.

You must ensure that the Application Instances to upgrade have been stopped prior to initiating the upgrade with this function.

The upgrade process will apply any new database migrations to each of the targeted Instances and then return a list of each Instance's upgrade status. Once upgraded, the Instances many be restarted; this process does not attempt to restart Instances.

Parameters

instances - Identifies the Application Instances to upgrade.

startup_options - The startup options which provide database connection information for the Instance that is to be upgraded. See MscmpSystOptions for more. This value is required.

opts - Miscellaneous options which the specific Application implementation of this Behaviour may define for it's own purposes.