Msdata.SystApplicationContexts (mscmp_syst_instance v0.1.0)

Defines the data structure of the Application Context.

Applications are written with certain security and connection characteristics in mind which correlate to database roles used by the application for establishing connections. This data type defines the Datastore contexts the application is expecting so that Instance records can be validated against the application expectations.

Defined in MscmpSystInstance.

Types

t()

@type t() :: %Msdata.SystApplicationContexts{
  __meta__: Ecto.Schema.Metadata.t(),
  application:
    Msdata.SystApplications.t() | Ecto.Association.NotLoaded.t() | nil,
  application_id: Ecto.UUID.t() | nil,
  database_owner_context: boolean() | nil,
  description: String.t() | nil,
  diag_role_created: String.t() | nil,
  diag_role_modified: String.t() | nil,
  diag_row_version: integer() | nil,
  diag_timestamp_created: DateTime.t() | nil,
  diag_timestamp_modified: DateTime.t() | nil,
  diag_update_count: integer() | nil,
  diag_wallclock_modified: DateTime.t() | nil,
  display_name: String.t() | nil,
  id: Ecto.UUID.t() | nil,
  internal_name: MscmpSystInstance.Types.application_context_name() | nil,
  login_context: boolean() | nil,
  start_context: boolean() | nil
}

Functions

insert_changeset(insert_params, opts \\ [])

@spec insert_changeset(
  MscmpSystInstance.Types.application_context_params(),
  Keyword.t()
) ::
  Ecto.Changeset.t()

Validates presented Application Context parameters for inserting a new Application Context record.

update_changeset(application_context, update_params, opts \\ [])

@spec update_changeset(
  t(),
  MscmpSystInstance.Types.application_context_params(),
  Keyword.t()
) :: Ecto.Changeset.t()

Validates update Application Context parameters for use in updating an existing Application Context record.