Msdata.SystApplications (mscmp_syst_instance v0.1.0)

Data structure describing the known applications for which instances may be hosted.

Defined in MscmpSystInstance.

Types

t()

@type t() :: %Msdata.SystApplications{
  __meta__: Ecto.Schema.Metadata.t(),
  application_contexts: term(),
  diag_role_created: String.t() | nil,
  diag_role_modified: String.t() | nil,
  diag_row_version: integer() | nil,
  diag_timestamp_created: DateTime.t() | nil,
  diag_timestamp_modified: DateTime.t() | nil,
  diag_update_count: integer() | nil,
  diag_wallclock_modified: DateTime.t() | nil,
  display_name: String.t() | nil,
  id: Ecto.UUID.t() | nil,
  instances: term(),
  internal_name: MscmpSystInstance.Types.application_name() | nil,
  syst_description: String.t() | nil
}

Functions

insert_changeset(insert_params, opts \\ [])

@spec insert_changeset(MscmpSystInstance.Types.application_params(), Keyword.t()) ::
  Ecto.Changeset.t()

Validates presented Application parameters for inserting a new Application record.

Parameters

insert_params: A map containing the parameters for creating a new Application.

opts: Optional keyword list of validation options.

Options

:min_internal_name_length (non_neg_integer/0) - Minimum length for internal names The default value is 6.

:max_internal_name_length (pos_integer/0) - Maximum length for internal names The default value is 64.

:min_display_name_length (non_neg_integer/0) - Minimum length for display names The default value is 6.

:max_display_name_length (pos_integer/0) - Maximum length for display names The default value is 64.

Returns

An Ecto.Changeset struct representing the validation result.

update_changeset(application, update_params, opts \\ [])

@spec update_changeset(t(), MscmpSystInstance.Types.application_params(), Keyword.t()) ::
  Ecto.Changeset.t()

Validates update Application parameters for use in updating an existing Application record.

Parameters

application: The existing Application struct to be updated.

update_params: A map containing the parameters for updating the Application.

opts: Optional keyword list of validation options.

Options

:min_internal_name_length (non_neg_integer/0) - Minimum length for internal names The default value is 6.

:max_internal_name_length (pos_integer/0) - Maximum length for internal names The default value is 64.

:min_display_name_length (non_neg_integer/0) - Minimum length for display names The default value is 6.

:max_display_name_length (pos_integer/0) - Maximum length for display names The default value is 64.

Returns

An Ecto.Changeset struct representing the validation result.