Msdata.SystApplications (mscmp_syst_instance v0.1.0)

Data structure describing the known applications for which instances may be hosted.

Defined in MscmpSystInstance.

Types

t()

@type t() :: %Msdata.SystApplications{
  __meta__: Ecto.Schema.Metadata.t(),
  application_contexts: term(),
  diag_role_created: String.t() | nil,
  diag_role_modified: String.t() | nil,
  diag_row_version: integer() | nil,
  diag_timestamp_created: DateTime.t() | nil,
  diag_timestamp_modified: DateTime.t() | nil,
  diag_update_count: integer() | nil,
  diag_wallclock_modified: DateTime.t() | nil,
  display_name: String.t() | nil,
  id: Ecto.UUID.t() | nil,
  instances: term(),
  internal_name: MscmpSystInstance.Types.application_name() | nil,
  syst_description: String.t() | nil
}

Functions

insert_changeset(insert_params, opts \\ [])

@spec insert_changeset(MscmpSystInstance.Types.application_params(), Keyword.t()) ::
  Ecto.Changeset.t()

Validates presented Application parameters for inserting a new Application record.

update_changeset(application, update_params, opts \\ [])

@spec update_changeset(t(), MscmpSystInstance.Types.application_params(), Keyword.t()) ::
  Ecto.Changeset.t()

Validates update Application parameters for use in updating an existing Application record.